refactor: isolate Navigation component into a dedicated folder
